OJ-2331: Move configuration helper to utils
