+++ test-resources/headless-core-stub/lambdas/callback/src/services/private-key-jwt-helper.ts
@@ -1,5 +1,5 @@
 import { v4 as uuidv4 } from "uuid";
-import { signJwt } from "../../../utils/crypto/signer";
+import { signJwt } from "../../../../utils/src/crypto/signer";
 import { JWK, JWTPayload } from "jose";
 export const generatePrivateJwtParams = async (
     clientId: string,
