+++ test-resources/headless-core-stub/lambdas/callback/tsconfig.json
@@ -15,5 +15,5 @@
         "noImplicitAny": true
     },
     "exclude": ["node_modules", "build", "coverage"],
-    "include": ["src", "tests", "../utils/**/*.ts"]
+    "include": ["src", "tests", "../../utils/**/*.ts"]
 }
