+++ test-resources/headless-core-stub/lambdas/start/tsconfig.json
@@ -15,5 +15,5 @@
         "noImplicitAny": true
     },
     "exclude": ["node_modules", "build", "coverage"],
-    "include": ["src", "tests"]
+    "include": ["src", "tests", "../../utils/**/*.ts"]
 }
