+++ test-resources/headless-core-stub/utils/jest.config.ts
@@ -1,7 +1,8 @@
 import type { Config } from "jest";
-import baseConfig from "../../../jest.config.base";
+import baseConfig from "../../jest.config.base";
 
 export default {
     ...baseConfig,
+    setupFiles: ["<rootDir>/setEnvVars.js"],
     displayName: "headless-core-stub/lambdas/callback",
 } satisfies Config;
