+++ test-resources/headless-core-stub/utils/setEnvVars.js
