+++ test-resources/headless-core-stub/utils/src/crypto/signer.ts
