+++ test-resources/jest.config.ts
@@ -1,5 +1,9 @@
 import type { Config } from "jest";
 
 export default {
-    projects: ["audit-events-test-harness/lambdas/jest.config.ts", "headless-core-stub/lambdas/*/jest.config.ts"],
+    projects: [
+        "audit-events-test-harness/lambdas/jest.config.ts",
+        "headless-core-stub/lambdas/*/jest.config.ts",
+        "headless-core-stub/utils/jest.config.ts",
+    ],
 } satisfies Config;
